Separate PlayerServiceTests
