+++ ShapeDungeon.Tests/ServiceTests/Players/PlayerGetServiceTests.cs
@@ -1,8 +1,6 @@
 ﻿#nullable disable
 using Moq;
 using NUnit.Framework;
-using ShapeDungeon.Data;
-using ShapeDungeon.DTOs.Players;
 using ShapeDungeon.Entities;
 using ShapeDungeon.Interfaces.Services.Players;
 using ShapeDungeon.Repos;
@@ -13,54 +11,16 @@ using System.Threading.Tasks;
 
 namespace ShapeDungeon.Tests.ServiceTests.Players
 {
-    internal class PlayerServiceTests
+    internal class PlayerGetServiceTests
     {
         private Mock<IPlayerRepository> _repoMock;
-        private Mock<IUnitOfWork> _unitOfWorkMock;
-        private IPlayerService _service;
+        private IPlayerGetService _service;
 
         [SetUp]
         public void Test_Initialize()
         {
             _repoMock = new Mock<IPlayerRepository>();
-            _unitOfWorkMock = new Mock<IUnitOfWork>();
-            _service = new PlayerService(_repoMock.Object, _unitOfWorkMock.Object);
-        }
-
-        [Test]
-        public async Task CreatePlayer_WithUniqueName_CreatesNewPlayer()
-        {
-            // Arrange
-            var uniqueName = "Minecraft Steve Lvl.64";
-            var playerDto = new PlayerDto { Name = uniqueName };
-
-            _repoMock
-                .Setup(x => x.DoesNameExist(playerDto.Name))
-                .ReturnsAsync(false);
-
-            // Act
-            var isCreated = await _service.CreatePlayerAsync(playerDto);
-
-            // Assert
-            Assert.IsTrue(isCreated);
-        }
-
-        [Test]
-        public async Task CreatePlayer_WithExistingName_DoesNotCreateNewPlayer()
-        {
-            // Arrange
-            var existingName = "Minecraft Steve Lvl.64";
-            var playerDto = new PlayerDto { Name = existingName };
-
-            _repoMock
-                .Setup(x => x.DoesNameExist(playerDto.Name))
-                .ReturnsAsync(true);
-
-            // Act
-            var isCreated = await _service.CreatePlayerAsync(playerDto);
-
-            // Assert
-            Assert.IsFalse(isCreated);
+            _service = new PlayerGetService(_repoMock.Object);
         }
 
         [Test]
